requirements.txt
# Core multi-agent framework
autogen-core>=0.4.0
autogen-ext[openai]>=0.4.0

# LLM / embedding clients
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0

# Knowledge base / document processing
python-docx>=1.1.0
chromadb>=0.5.0

# Data / validation
pydantic>=2.0.0
